Add support for multiple chain types in the configuration (#2228)
* Move `ChainEndpoint` trait into its own module

* Add support for multiple chain types

* Add changelog entry

* Remove `type` entry from the default config

* Update relayer-cli/src/cli_utils.rs


* Ignore case and dashes when deserializing `ChainType`

* Fix test
