+++ .changelog/unreleased/features/ibc-relayer/2240-chain-types.md
@@ -1,0 +1,3 @@
+- Add preliminary support for multiple chain types, which can be specified in
+  the chain configuration. At the moment only the `CosmosSdk` chain type is
+  supported. ([#2240](https://github.com/informalsystems/ibc-rs/issues/2240))
