+++ relayer-cli/src/cli_utils.rs
@@ -1,16 +1,17 @@
 use alloc::sync::Arc;
+
 use tokio::runtime::Runtime as TokioRuntime;
 
 use ibc::core::ics02_client::client_state::ClientState;
 use ibc::core::ics24_host::identifier::{ChainId, ChannelId, PortId};
-use ibc_relayer::chain::counterparty::{channel_connection_client, ChannelConnectionClient};
+
 use ibc_relayer::{
     chain::{
+        counterparty::{channel_connection_client, ChannelConnectionClient},
         handle::{BaseChainHandle, ChainHandle},
-        runtime::ChainRuntime,
-        CosmosSdkChain,
     },
     config::Config,
+    spawn,
 };
 
 use crate::error::Error;
@@ -55,20 +56,12 @@ pub fn spawn_chain_runtime(config: &Config, chain_id: &ChainId) -> Result<impl C
     spawn_chain_runtime_generic::<BaseChainHandle>(config, chain_id)
 }
 
-pub fn spawn_chain_runtime_generic<Chain: ChainHandle>(
+pub fn spawn_chain_runtime_generic<Handle: ChainHandle>(
     config: &Config,
     chain_id: &ChainId,
-) -> Result<Chain, Error> {
-    let chain_config = config
-        .find_chain(chain_id)
-        .cloned()
-        .ok_or_else(|| Error::missing_chain_config(chain_id.clone()))?;
-
+) -> Result<Handle, Error> {
     let rt = Arc::new(TokioRuntime::new().unwrap());
-    let handle =
-        ChainRuntime::<CosmosSdkChain>::spawn::<Chain>(chain_config, rt).map_err(Error::relayer)?;
-
-    Ok(handle)
+    spawn::spawn_chain_runtime(config, chain_id, rt).map_err(Error::spawn)
 }
 
 /// Spawns a chain runtime for specified chain identifier, queries the counterparty chain associated
