+++ relayer-cli/src/commands/query/client.rs
@@ -1,13 +1,12 @@
-use alloc::sync::Arc;
-
 use abscissa_core::clap::Parser;
 use abscissa_core::{Command, Runnable};
+use tracing::debug;
+
+use ibc_relayer::chain::handle::ChainHandle;
 use ibc_relayer::chain::requests::{
     PageRequest, QueryClientConnectionsRequest, QueryClientStateRequest,
     QueryConsensusStateRequest, QueryConsensusStatesRequest,
 };
-use tokio::runtime::Runtime as TokioRuntime;
-use tracing::debug;
 
 use ibc::core::ics02_client::client_consensus::QueryClientEventRequest;
 use ibc::core::ics02_client::client_state::ClientState;
@@ -16,10 +15,9 @@ use ibc::core::ics24_host::identifier::ClientId;
 use ibc::events::WithBlockDataType;
 use ibc::query::QueryTxRequest;
 use ibc::Height;
-use ibc_relayer::chain::ChainEndpoint;
-use ibc_relayer::chain::CosmosSdkChain;
 
 use crate::application::app_config;
+use crate::cli_utils::spawn_chain_runtime;
 use crate::conclude::{exit_with_unrecoverable_error, Output};
 
 /// Query client state command
@@ -41,18 +39,9 @@ impl Runnable for QueryClientStateCmd {
     fn run(&self) {
         let config = app_config();
 
-        let chain_config = match config.find_chain(&self.chain_id) {
-            None => Output::error(format!(
-                "chain '{}' not found in configuration file",
-                self.chain_id
-            ))
-            .exit(),
-            Some(chain_config) => chain_config,
-        };
-
-        let rt = Arc::new(TokioRuntime::new().unwrap());
-        let chain = CosmosSdkChain::bootstrap(chain_config.clone(), rt)
+        let chain = spawn_chain_runtime(&config, &self.chain_id)
             .unwrap_or_else(exit_with_unrecoverable_error);
+
         let height = ibc::Height::new(chain.id().version(), self.height.unwrap_or(0_u64));
 
         match chain.query_client_state(QueryClientStateRequest {
@@ -98,19 +87,9 @@ impl Runnable for QueryClientConsensusCmd {
     fn run(&self) {
         let config = app_config();
 
-        let chain_config = match config.find_chain(&self.chain_id) {
-            None => Output::error(format!(
-                "chain '{}' not found in configuration file",
-                self.chain_id
-            ))
-            .exit(),
-            Some(chain_config) => chain_config,
-        };
-
         debug!("Options: {:?}", self);
 
-        let rt = Arc::new(TokioRuntime::new().unwrap());
-        let chain = CosmosSdkChain::bootstrap(chain_config.clone(), rt)
+        let chain = spawn_chain_runtime(&config, &self.chain_id)
             .unwrap_or_else(exit_with_unrecoverable_error);
 
         let counterparty_chain = match chain.query_client_state(QueryClientStateRequest {
@@ -184,19 +163,9 @@ impl Runnable for QueryClientHeaderCmd {
     fn run(&self) {
         let config = app_config();
 
-        let chain_config = match config.find_chain(&self.chain_id) {
-            None => Output::error(format!(
-                "chain '{}' not found in configuration file",
-                self.chain_id
-            ))
-            .exit(),
-            Some(chain_config) => chain_config,
-        };
-
         debug!("Options: {:?}", self);
 
-        let rt = Arc::new(TokioRuntime::new().unwrap());
-        let chain = CosmosSdkChain::bootstrap(chain_config.clone(), rt)
+        let chain = spawn_chain_runtime(&config, &self.chain_id)
             .unwrap_or_else(exit_with_unrecoverable_error);
 
         let counterparty_chain = match chain.query_client_state(QueryClientStateRequest {
@@ -213,6 +182,7 @@ impl Runnable for QueryClientHeaderCmd {
 
         let consensus_height =
             ibc::Height::new(counterparty_chain.version(), self.consensus_height);
+
         let height = ibc::Height::new(chain.id().version(), self.height.unwrap_or(0_u64));
 
         let res = chain.query_txs(QueryTxRequest::Client(QueryClientEventRequest {
@@ -251,19 +221,9 @@ impl Runnable for QueryClientConnectionsCmd {
     fn run(&self) {
         let config = app_config();
 
-        let chain_config = match config.find_chain(&self.chain_id) {
-            None => Output::error(format!(
-                "chain '{}' not found in configuration file",
-                self.chain_id
-            ))
-            .exit(),
-            Some(chain_config) => chain_config,
-        };
-
         debug!("Options: {:?}", self);
 
-        let rt = Arc::new(TokioRuntime::new().unwrap());
-        let chain = CosmosSdkChain::bootstrap(chain_config.clone(), rt)
+        let chain = spawn_chain_runtime(&config, &self.chain_id)
             .unwrap_or_else(exit_with_unrecoverable_error);
 
         let res = chain.query_client_connections(QueryClientConnectionsRequest {
