+++ relayer/src/chain/cosmos/query.rs
@@ -2,7 +2,7 @@ use http::uri::Uri;
 use ibc::core::ics02_client::client_consensus::QueryClientEventRequest;
 use ibc::core::ics04_channel::channel::QueryPacketEventDataRequest;
 use ibc::core::ics04_channel::packet::Sequence;
-use ibc::core::ics23_commitment::merkle::convert_tm_to_ics_merkle_proof;
+use ibc::core::ics23_commitment::merkle::{convert_tm_to_ics_merkle_proof, MerkleProof};
 use ibc::core::ics24_host::identifier::ChainId;
 use ibc::query::QueryTxHash;
 use ibc_proto::cosmos::base::tendermint::v1beta1::service_client::ServiceClient;
@@ -13,13 +13,20 @@ use tendermint_rpc::query::Query;
 use tendermint_rpc::{Client, HttpClient, Url};
 
 use crate::chain::cosmos::version::Specs;
-use crate::chain::QueryResponse;
 use crate::error::Error;
 
 pub mod account;
 pub mod balance;
 pub mod status;
 pub mod tx;
+
+/// Generic query response type
+#[derive(Clone, Debug, PartialEq)]
+pub struct QueryResponse {
+    pub value: Vec<u8>,
+    pub proof: Option<MerkleProof>,
+    pub height: Height,
+}
 
 pub fn packet_query(request: &QueryPacketEventDataRequest, seq: Sequence) -> Query {
     Query::eq(
