+++ relayer/src/chain/handle.rs
@@ -42,6 +42,7 @@ use crate::{
 
 use super::{
     client::ClientSettings,
+    endpoint::{ChainStatus, HealthCheck},
     requests::{
         QueryChannelClientStateRequest, QueryChannelRequest, QueryChannelsRequest,
         QueryClientConnectionsRequest, QueryClientStateRequest, QueryClientStatesRequest,
@@ -53,7 +54,6 @@ use super::{
     },
     tracking::TrackedMsgs,
 };
-use super::{ChainStatus, HealthCheck};
 
 mod base;
 mod cache;
