+++ relayer/src/chain/mock.rs
@@ -31,10 +31,10 @@ use ibc::Height;
 
 use crate::account::Balance;
 use crate::chain::client::ClientSettings;
+use crate::chain::endpoint::{ChainEndpoint, ChainStatus, HealthCheck};
 use crate::chain::requests::{
     QueryChannelClientStateRequest, QueryChannelRequest, QueryClientStatesRequest,
 };
-use crate::chain::{ChainEndpoint, ChainStatus};
 use crate::config::ChainConfig;
 use crate::error::Error;
 use crate::event::monitor::{EventReceiver, EventSender, TxMonitorCmd};
@@ -51,7 +51,6 @@ use super::requests::{
     QueryUpgradedClientStateRequest, QueryUpgradedConsensusStateRequest,
 };
 use super::tracking::TrackedMsgs;
-use super::HealthCheck;
 
 /// The representation of a mocked chain as the relayer sees it.
 /// The relayer runtime and the light client will engage with the MockChain to query/send tx; the
@@ -451,12 +450,16 @@ pub mod test_utils {
 
     use ibc::core::ics24_host::identifier::ChainId;
 
-    use crate::config::{AddressType, ChainConfig, GasPrice, PacketFilter};
+    use crate::{
+        chain::ChainType,
+        config::{AddressType, ChainConfig, GasPrice, PacketFilter},
+    };
 
     /// Returns a very minimal chain configuration, to be used in initializing `MockChain`s.
     pub fn get_basic_chain_config(id: &str) -> ChainConfig {
         ChainConfig {
             id: ChainId::from_str(id).unwrap(),
+            r#type: ChainType::Mock,
             rpc_addr: "http://127.0.0.1:26656".parse().unwrap(),
             grpc_addr: "http://127.0.0.1:9090".parse().unwrap(),
             websocket_addr: "ws://127.0.0.1:26656/websocket".parse().unwrap(),
