+++ relayer/src/config.rs
@@ -16,6 +16,7 @@ use ibc::core::ics23_commitment::specs::ProofSpecs;
 use ibc::core::ics24_host::identifier::{ChainId, ChannelId, PortId};
 use ibc::timestamp::ZERO_DURATION;
 
+use crate::chain::ChainType;
 use crate::config::types::{MaxMsgNum, MaxTxSize, Memo};
 use crate::keyring::Store;
 
@@ -44,6 +45,10 @@ impl fmt::Display for GasPrice {
 /// Defaults for various fields
 pub mod default {
     use super::*;
+
+    pub fn chain_type() -> ChainType {
+        ChainType::CosmosSdk
+    }
 
     pub fn tx_confirmation() -> bool {
         true
@@ -309,6 +314,8 @@ impl fmt::Display for AddressType {
 #[serde(deny_unknown_fields)]
 pub struct ChainConfig {
     pub id: ChainId,
+    #[serde(default = "default::chain_type")]
+    pub r#type: ChainType,
     pub rpc_addr: tendermint_rpc::Url,
     pub websocket_addr: tendermint_rpc::Url,
     pub grpc_addr: tendermint_rpc::Url,
