+++ relayer/src/upgrade_chain.rs
@@ -18,9 +18,9 @@ use ibc_proto::cosmos::upgrade::v1beta1::Plan;
 use ibc_proto::google::protobuf::Any;
 use ibc_proto::ibc::core::client::v1::UpgradeProposal;
 
+use crate::chain::handle::ChainHandle;
 use crate::chain::requests::QueryClientStateRequest;
 use crate::chain::tracking::TrackedMsgs;
-use crate::chain::{ChainEndpoint, CosmosSdkChain};
 use crate::config::ChainConfig;
 use crate::error::Error;
 
@@ -66,12 +66,12 @@ pub struct UpgradePlanOptions {
 }
 
 pub fn build_and_send_ibc_upgrade_proposal(
-    mut dst_chain: CosmosSdkChain, // the chain which will undergo an upgrade
-    src_chain: CosmosSdkChain, // the source chain; supplies a client state for building the upgrade plan
+    dst_chain: impl ChainHandle, // the chain which will undergo an upgrade
+    src_chain: impl ChainHandle, // the source chain; supplies a client state for building the upgrade plan
     opts: &UpgradePlanOptions,
 ) -> Result<TxHash, UpgradeChainError> {
     let upgrade_height = dst_chain
-        .query_chain_latest_height()
+        .query_latest_height() // FIXME(romac): Use query_chain_latest_height once added to ChainHandle
         .map_err(UpgradeChainError::query)?
         .add(opts.height_offset);
 
@@ -147,7 +147,7 @@ pub fn build_and_send_ibc_upgrade_proposal(
 
     let responses = dst_chain
         .send_messages_and_wait_check_tx(TrackedMsgs::new_single(any_msg, "upgrade"))
-        .map_err(|e| UpgradeChainError::submit(dst_chain.id().clone(), e))?;
+        .map_err(|e| UpgradeChainError::submit(dst_chain.id(), e))?;
 
     Ok(responses[0].hash)
 }
